Let's update the one announcement call, too
