+++ src/backend/twitch-api/resource/chat.ts
@@ -20,17 +20,28 @@ export async function sendAnnouncement(
 ): Promise<boolean> {
     const client: ApiClient = sendAsBot === true ? twitchApi.getBotClient() : twitchApi.getClient();
     const streamerUserId: number = accountAccess.getAccounts().streamer.userId;
-    const senderUserId: number = sendAsBot === true ?
+    let senderUserId: number = sendAsBot === true && accountAccess.getAccounts().bot?.userId != null ?
         accountAccess.getAccounts().bot.userId :
         streamerUserId;
 
-    try {
-        const announcement: HelixSendChatAnnoucementParams = {
-            message: message,
-            color: color
-        };
+    if (message?.length < 1) {
+        return;
+    }
 
-        await client.chat.sendAnnouncement(streamerUserId, senderUserId, announcement);
+    try {        
+        // split message into fragments so we don't exceed the max message length
+        const messageFragments = message.match(/[\s\S]{1,500}/g)
+            .map(mf => mf.trim())
+            .filter(mf => mf !== "");
+
+        for (const fragment of messageFragments) {
+            const announcement: HelixSendChatAnnoucementParams = {
+                message: fragment,
+                color: color
+            };
+    
+            await client.chat.sendAnnouncement(streamerUserId, senderUserId, announcement);
+        }
 
         return true;
     } catch (error) {
